requirements.txt
flask
line-bot-sdk
google-generativeai
gspread
oauth2client
gspread
google-auth
